Take vreg regclasses from the Function interface explicitly.
Currently, we infer this from the first def we see of a vreg, but this
is somewhat brittle, as I have found when debugging the integration of
regalloc2 with Cranelift. It would be better to just obtain the register
class for each vreg directly from the embedder. This is a small change
that should be a direct type lookup in most embedders.