Simplify validate_ssa
Visiting the blocks in any order where dominators are visited before
their descendents means that definition validity can be checked in a
single pass over the blocks. I used breadth-first order on the dominance
tree because I found that slightly easier to implement than depth-first
while avoiding recursion.

With that done, the `defined` vector is redundant with `vreg_def_inst`
and `vreg_def_blockparam`. But the latter two were actually only used to
identify which block each value was defined in. So it's simpler to just
record the definition block for every value, without distinguishing
between blockparams and instructions. So instead of three vectors with
`num_vregs` entries each, there's only one.